# docs/build-customize-contribute/swagger.yaml
swagger: '2.0'
info:
  title: Harbor API
  description: These APIs provide services for manipulating Harbor project.
  version: 1.9.0
host: localhost
schemes:
  - http
  - https
basePath: /api
produces:
  - application/json
  - text/plain
consumes:
  - application/json
securityDefinitions:
  basicAuth:
    type: basic
security:
  - basicAuth: []
paths:
...
  '/repositories/{repo_name}/tags/{tag}':
    get:
      summary: Get the tag of the repository.
      description: |
        This endpoint aims to retrieve the tag of the repository. If deployed with Notary, the signature property of response represents whether the image is singed or not. If the property is null, the image is unsigned.
      parameters:
        - name: repo_name
          in: path
          type: string
          required: true
          description: Relevant repository name.
        - name: tag
          in: path
          type: string
          required: true
          description: Tag of the repository.
      tags:
        - Products
      responses:
        '200':
          description: Get tag successfully.
          schema:
            $ref: '#/definitions/DetailedTag'
        '500':
          description: Unexpected internal errors.
    delete:
      summary: Delete a tag in a repository.
      description: |
        This endpoint let user delete tags with repo name and tag.
      parameters:
        - name: repo_name
          in: path
          type: string
          required: true
          description: The name of repository which will be deleted.
        - name: tag
          in: path
          type: string
          required: true
          description: Tag of a repository.
      tags:
        - Products
      responses:
        '200':
          description: Delete tag successfully.
        '400':
          description: Invalid repo_name.
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '404':
          description: Repository or tag not found.
...
definitions:
...
  DetailedTag:
    type: object
    properties:
      digest:
        type: string
        description: The digest of the tag.
      name:
        type: string
        description: The name of the tag.
      size:
        type: integer
        description: The size of the image.
      architecture:
        type: string
        description: The architecture of the image.
      os:
        type: string
        description: The os of the image.
      docker_version:
        type: string
        description: The version of docker which builds the image.
      author:
        type: string
        description: The author of the image.
      created:
        type: string
        description: The build time of the image.
      signature:
        type: object
        description: 'The signature of image, defined by RepoSignature. If it is null, the image is unsigned.'
      scan_overview:
        type: object
        description: The overview of the scan result.  This is an optional property.
        properties:
          digest:
            type: string
            description: The digest of the image.
          scan_status:
            type: string
            description: 'The status of the scan job, it can be "pending", "running", "finished", "error".'
          job_id:
            type: integer
            description: The ID of the job on jobservice to scan the image.
          severity:
            type: integer
            description: '0-Not scanned, 1-Negligible, 2-Unknown, 3-Low, 4-Medium, 5-High'
          details_key:
            type: string
            description: 'The top layer name of this image in Clair, this is for calling Clair API to get the vulnerability list of this image.'
          components:
            type: object
            description: The components overview of the image.
            properties:
              total:
                type: integer
                description: Total number of the components in this image.
              summary:
                description: List of number of components of different severities.
                type: array
                items:
                  $ref: '#/definitions/ComponentOverviewEntry'
      labels:
        type: array
        description: The label list.
        items:
          $ref: '#/definitions/Label'
  ComponentOverviewEntry:
    type: object
    properties:
      severity:
        type: integer
        description: '1-None/Negligible, 2-Unknown, 3-Low, 4-Medium, 5-High'
      count:
        type: integer
        description: number of the components with certain severity.
...
  Label:
    type: object
    properties:
      id:
        type: integer
        description: The ID of label.
      name:
        type: string
        description: The name of label.
      description:
        type: string
        description: The description of label.
      color:
        type: string
        description: The color of label.
      scope:
        type: string
        description: 'The scope of label, g for global labels and p for project labels.'
      project_id:
        type: integer
        description: The project ID if the label is a project label.
      creation_time:
        type: string
        description: The creation time of label.
      update_time:
        type: string
        description: The update time of label.
      deleted:
        type: boolean
        description: The label is deleted or not.
...